Add another version of the example app to demo auth/token exchange
