+++ Example/Latest/components/Authenticate.js
@@ -1,0 +1,129 @@
+import React, { Component } from 'react';
+import { UIManager, LayoutAnimation, Alert, ScrollView } from 'react-native';
+import { authorize, refresh, revoke } from 'react-native-app-auth';
+import { Page, Button, ButtonContainer, Form, Heading } from './';
+
+UIManager.setLayoutAnimationEnabledExperimental &&
+  UIManager.setLayoutAnimationEnabledExperimental(true);
+
+type State = {
+  hasLoggedInOnce: boolean,
+  accessToken: ?string,
+  accessTokenExpirationDate: ?string,
+  refreshToken: ?string
+};
+
+const config = {
+  issuer: 'https://demo.identityserver.io',
+  clientId: 'native.code',
+  redirectUrl: 'io.identityserver.demo:/oauthredirect',
+  additionalParameters: {},
+  scopes: ['openid', 'profile', 'email', 'offline_access']
+
+  // serviceConfiguration: {
+  //   authorizationEndpoint: 'https://demo.identityserver.io/connect/authorize',
+  //   tokenEndpoint: 'https://demo.identityserver.io/connect/token',
+  //   revocationEndpoint: 'https://demo.identityserver.io/connect/revoke'
+  // }
+};
+
+export default class App extends Component<{}, State> {
+  state = {
+    hasLoggedInOnce: false,
+    accessToken: '',
+    accessTokenExpirationDate: '',
+    refreshToken: ''
+  };
+
+  animateState(nextState: $Shape<State>, delay: number = 0) {
+    setTimeout(() => {
+      this.setState(() => {
+        LayoutAnimation.easeInEaseOut();
+        return nextState;
+      });
+    }, delay);
+  }
+
+  authorize = async () => {
+    try {
+      const authState = await authorize(config);
+
+      this.animateState(
+        {
+          hasLoggedInOnce: true,
+          accessToken: authState.accessToken,
+          accessTokenExpirationDate: authState.accessTokenExpirationDate,
+          refreshToken: authState.refreshToken,
+          scopes: authState.scopes
+        },
+        500
+      );
+    } catch (error) {
+      Alert.alert('Failed to log in', error.message);
+    }
+  };
+
+  refresh = async () => {
+    try {
+      const authState = await refresh(config, {
+        refreshToken: this.state.refreshToken
+      });
+
+      this.animateState({
+        accessToken: authState.accessToken || this.state.accessToken,
+        accessTokenExpirationDate:
+          authState.accessTokenExpirationDate || this.state.accessTokenExpirationDate,
+        refreshToken: authState.refreshToken || this.state.refreshToken
+      });
+    } catch (error) {
+      Alert.alert('Failed to refresh token', error.message);
+    }
+  };
+
+  revoke = async () => {
+    try {
+      await revoke(config, {
+        tokenToRevoke: this.state.accessToken,
+        sendClientId: true
+      });
+      this.animateState({
+        accessToken: '',
+        accessTokenExpirationDate: '',
+        refreshToken: ''
+      });
+    } catch (error) {
+      Alert.alert('Failed to revoke token', error.message);
+    }
+  };
+
+  render() {
+    const { state } = this;
+    return (
+      <>
+        <ScrollView>
+          {!!state.accessToken ? (
+            <Form>
+              <Form.Label>accessToken</Form.Label>
+              <Form.Value>{state.accessToken}</Form.Value>
+              <Form.Label>accessTokenExpirationDate</Form.Label>
+              <Form.Value>{state.accessTokenExpirationDate}</Form.Value>
+              <Form.Label>refreshToken</Form.Label>
+              <Form.Value>{state.refreshToken}</Form.Value>
+              <Form.Label>Scopes</Form.Label>
+              <Form.Value>{state.scopes.join(', ')}</Form.Value>
+            </Form>
+          ) : (
+            <Heading>{state.hasLoggedInOnce ? 'Goodbye.' : 'Hello, stranger.'}</Heading>
+          )}
+        </ScrollView>
+        <ButtonContainer>
+          {!state.accessToken && (
+            <Button onPress={this.authorize} text="Authorize" color="#DA2536" />
+          )}
+          {!!state.refreshToken && <Button onPress={this.refresh} text="Refresh" color="#24C2CB" />}
+          {!!state.accessToken && <Button onPress={this.revoke} text="Revoke" color="#EF525B" />}
+        </ButtonContainer>
+      </>
+    );
+  }
+}
