+++ Example/Latest/components/index.js
@@ -5,3 +5,5 @@ export { default as ButtonContainer } from './ButtonContainer';
 export { default as Form } from './Form';
 export { default as Heading } from './Heading';
 export { default as Page } from './Page';
+export { default as Advanced } from './Advanced';
+export { default as Authenticate } from './Authenticate';
