+++ index.js
@@ -1,5 +1,5 @@
-export { deafault as onlyTokenExchange } from './src/only-token-exchange';
-export { deafault as onlyAuthorize } from './src/only-authorize';
-export { deafault as authorize } from './src/authorize';
-export { deafault as revoke } from './src/revoke';
-export { deafault as refresh } from './src/refresh';
+export { default as onlyTokenExchange } from './src/only-token-exchange';
+export { default as onlyAuthorize } from './src/only-authorize';
+export { default as authorize } from './src/authorize';
+export { default as revoke } from './src/revoke';
+export { default as refresh } from './src/refresh';
